requirements.txt
pyttsx3==2.99
pywhatkit==5.4
SpeechRecognition==3.14.3
google-generativeai==0.4.1
protobuf==4.25.8
pyaudio==0.2.14
google-auth==2.40.3